rewrite of environment
