+++ aspnetcore/fundamentals/environments/sample/About.cshtml
@@ -1,0 +1,21 @@
+﻿@page
+@model AboutModel
+@{
+    ViewData["Title"] = "About";
+}
+<h2>@ViewData["Title"]</h2>
+<h3>@Model.Message</h3>
+
+<p>Use this area to provide additional information.</p>
+
+<environment include="Development">
+    <div>&lt;environment include="Development"&gt;</div>
+</environment>
+<environment exclude="Development">
+    <div>&lt;environment exclude="Development"&gt;</div>
+</environment>
+<environment include="Staging,Development,Staging_2">
+    <div>
+        &lt;environment include="Staging,Development,Staging_2"&gt;
+    </div>
+</environment>
