--- aspnetcore/fundamentals/environments/sample/Environments.sln
@@ -1,32 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 14
-VisualStudioVersion = 14.0.22823.1
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "src", "src", "{73B4581B-78B1-4D1F-86AA-915D09EBB68F}"
-EndProject
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{B18D8638-0CBD-4BB6-99AA-569E3747E907}"
-	ProjectSection(SolutionItems) = preProject
-		global.json = global.json
-	EndProjectSection
-EndProject
-Project("{8BB2217D-0F2D-49D1-97BC-3654ED321F3B}") = "Environments", "src\Environments\Environments.xproj", "{9D14116C-527D-418A-B937-C826D2F6519E}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{9D14116C-527D-418A-B937-C826D2F6519E}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{9D14116C-527D-418A-B937-C826D2F6519E}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{9D14116C-527D-418A-B937-C826D2F6519E}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{9D14116C-527D-418A-B937-C826D2F6519E}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(NestedProjects) = preSolution
-		{9D14116C-527D-418A-B937-C826D2F6519E} = {73B4581B-78B1-4D1F-86AA-915D09EBB68F}
-	EndGlobalSection
-EndGlobal
