--- aspnetcore/fundamentals/environments/sample/global.json
@@ -1,3 +1,0 @@
-﻿{
-    "projects": [ "src", "test" ]
-}
