--- aspnetcore/fundamentals/environments/sample/src/Environments/Program.cs
@@ -1,21 +1,0 @@
-﻿using Microsoft.AspNetCore.Hosting;
-using System.IO;
-using System.Reflection;
-
-namespace Environments
-{
-    public class Program
-    {
-        public static void Main(string[] args)
-        {
-            var host = new WebHostBuilder()
-                .UseKestrel()
-                .UseContentRoot(Directory.GetCurrentDirectory())
-                .UseIISIntegration()
-                .UseStartup<Startup>()
-                .Build();
-
-            host.Run();
-        }
-    }
-}
