--- aspnetcore/fundamentals/environments/sample/src/Environments/Startup.cs
@@ -1,40 +1,0 @@
-﻿using Microsoft.AspNetCore.Builder;
-using Microsoft.AspNetCore.Hosting;
-using Microsoft.AspNetCore.Http;
-using Microsoft.Extensions.DependencyInjection;
-
-namespace Environments
-{
-    public class Startup
-    {
-        // For more information on how to configure your application, visit http://go.microsoft.com/fwlink/?LinkID=398940
-        public void ConfigureServices(IServiceCollection services)
-        {
-        }
-
-        // When editing this document, do not modify the STARTSNIP/ENDSNIP comments.
-        // They are used to locate where in this document a specific code snippet is located
-        // If you are using the sample in your own application, feel free to remove the comments.
-        // STARTSNIP:error-handling
-        public void Configure(IApplicationBuilder app, IHostingEnvironment env)
-        {
-            if (env.IsDevelopment())
-            {
-                app.UseDeveloperExceptionPage();
-                app.UseDatabaseErrorPage();
-                app.UseBrowserLink();
-            }
-            else
-            {
-                app.UseExceptionHandler("/Home/Error");
-            }
-            // ...
-            // ENDSNIP:error-handling
-
-            app.Run(async (context) =>
-            {
-                await context.Response.WriteAsync("Hello World!");
-            });
-        }
-    }
-}
