--- aspnetcore/fundamentals/environments/sample/src/Environments/wwwroot/placeholder.txt
@@ -1,1 +1,0 @@
-﻿﻿Placefolder file to ensure the existing of wwwroot folder.
